Fixed continent bug (needs editing)
Something went wrong in displaying "instance of" relations, it did not get 'value' == "instance of" but instead 'value'  == www.wikidata/etc/etc/P31, which doesn't work with the word-matching that was implemented. There's now matching between this url and Is.
This was tested on:
Is France a country (yes)
Is France a continent (no)
Is Asia a country (no)
Is Asia a continent (yes)

Also, I stopped nounifying for these, since nounifying goes on for too long, which means that simple questions like "Is Paris the capital of Germany" end up not as "no", but as an infinite nounifying loop.